+++ ui/src/app/core/components/required-indicator/required-indicator.component.ts
@@ -1,0 +1,9 @@
+import { Component } from '@angular/core'
+
+@Component({
+  selector: 'app-required-indicator',
+  templateUrl: './required-indicator.component.html',
+  styleUrls: ['./required-indicator.component.scss'],
+  standalone: true,
+})
+export class RequiredIndicatorComponent {}